[example] refactor code
